refactor: implement search in abstract class
in order to reduce code duplication.
