+++ src/Api/Adapter/TeamSiteAdapter.php
@@ -10,11 +10,12 @@ use Omeka\Api\Exception;
 use Omeka\Api\Request;
 use Omeka\Api\Response;
 use Omeka\Entity\EntityInterface;
+use Omeka\Entity\Site;
 use Omeka\Stdlib\ErrorStore;
 use Teams\Entity\TeamSite;
 use Teams\Api\Representation\TeamSiteRepresentation;
 
-class TeamSiteAdapter extends AbstractEntityAdapter
+class TeamSiteAdapter extends AbstractTeamEntityAdapter
 {
     protected $sortFields = [
         'site' => 'site',
@@ -103,174 +104,52 @@ class TeamSiteAdapter extends AbstractEntityAdapter
         }
     }
 
-    public function search(Request $request)
-    {
-        $search_fields = array();
-        $group_by = 'team'; //default order by
-        $query = $request->getContent();
-
-        if ( array_key_exists('team', $query) ) {
-            $search_fields['team'] = $query['team'];
-            $group_by = 'site';
-        } elseif (array_key_exists('site', $query)) {
-            $search_fields['site'] = $query['site'];
-        } else {
-            throw new Exception\BadRequestException(sprintf(
-                $this->getTranslator()->translate('%1$s entity requires team or site search criteria'),
-                $this->getEntityClass()
-            ));
-        }
-
-        // Set default query parameters
-        if (!isset($query['page'])) {
-            $query['page'] = null;
-        }
-        if (!isset($query['per_page'])) {
-            $query['per_page'] = null;
-        }
-        if (!isset($query['limit'])) {
-            $query['limit'] = null;
-        }
-        if (!isset($query['offset'])) {
-            $query['offset'] = null;
-        }
-        if (!isset($query['sort_by'])) {
-            $query['sort_by'] = null;
-        }
-        if (isset($query['sort_order'])
-            && in_array(strtoupper($query['sort_order']), ['ASC', 'DESC'])
-        ) {
-            $query['sort_order'] = strtoupper($query['sort_order']);
-        } else {
-            $query['sort_order'] = 'ASC';
-        }
-        if (!isset($query['return_scalar'])) {
-            $query['return_scalar'] = null;
-        }
-
-        // Begin building the search query.
-        $entityClass = $this->getEntityClass();
-
-        $this->index = 0;
-        $qb = $this->getEntityManager()
-            ->createQueryBuilder()
-            ->select('omeka_root')
-            ->from($entityClass, 'omeka_root');
-
-        foreach ($search_fields as $field => $value) {
-            $qb->andWhere($qb->expr()->eq(
-                "omeka_root.$field",
-                $this->createNamedParameter($qb, $value)
-            ));
-        }
-        $this->buildBaseQuery($qb, $query);
-        $this->buildQuery($qb, $query);
-        $qb->groupBy("omeka_root." . $group_by);
-
-
-
-        // Add the LIMIT clause.
-        $this->limitQuery($qb, $query);
-
-        // Before adding the ORDER BY clause, set a paginator responsible for
-        // getting the total count. This optimization excludes the ORDER BY
-        // clause from the count query, greatly speeding up response time.
-        $countQb = clone $qb;
-        $countQb->select('1')->resetDQLPart('orderBy');
-        $countPaginator = new Paginator($countQb, false);
-
-        // Add the ORDER BY clause. Always sort by entity ID in addition to any
-        // sorting the adapters add.
-        $this->sortQuery($qb, $query);
-        $qb->addOrderBy("omeka_root.team", $query['sort_order']);
-
-        $scalarField = $request->getOption('returnScalar');
-        if (!$scalarField && $query['return_scalar']) {
-            if (!array_key_exists($query['return_scalar'], $this->scalarFields)) {
-                throw new Exception\BadRequestException(sprintf(
-                    $this->getTranslator()->translate('The "%1$s" field is not available in the %2$s adapter class.'),
-                    $query['return_scalar'], get_class($this)
-                ));
-            }
-            // The return_scalar passed in the query is valid. Note that we must
-            // set returnScalar to the request so the API manager skips validation.
-            $scalarField = $query['return_scalar'];
-            $request->setOption('returnScalar', $scalarField);
-        }
-        if ($scalarField) {
-            $classMetadata = $this->getEntityManager()->getClassMetadata($entityClass);
-            $fieldNames = $classMetadata->getFieldNames();
-            if (!in_array($scalarField, $fieldNames)) {
-                $associationNames = $classMetadata->getAssociationNames();
-                if (!in_array($scalarField, $associationNames)) {
-                    throw new Exception\BadRequestException(sprintf(
-                        $this->getTranslator()->translate('The "%1$s" field is not available in the %2$s entity class.'),
-                        $scalarField, $entityClass
-                    ));
-                }
-                $qb->select(["IDENTITY(omeka_root.team) AS team, IDENTITY(omeka_root.site) as site"]);
-            } else {
-                $qb->select(['omeka_root.id', 'omeka_root.' . $scalarField]);
-            }
-            $content = array_column($qb->getQuery()->getScalarResult(), $scalarField, $scalarField);
-            $response = new Response($content);
-            $response->setTotalResults($countPaginator->count());
-            return $response;
-        }
-
-
-        $paginator = new Paginator($qb, false);
-        $entities = [];
-        // Don't make the request if the LIMIT is set to zero. Useful if the
-        // only information needed is total results.
-        if ($qb->getMaxResults() || null === $qb->getMaxResults()) {
-            foreach ($paginator as $entity) {
-                if (is_array($entity)) {
-                    // Remove non-entity columns added to the SELECT. You can use
-                    // "AS HIDDEN {alias}" to avoid this condition.
-                    $entity = $entity[0];
-                }
-                $entities[] = $entity;
-            }
-        }
-
-        $response = new Response($entities);
-        $response->setTotalResults($countPaginator->count());
-        return $response;
-    }
-
     public function read(Request $request)
     {
-        AbstractAdapter::read($request);
+        AbstractTeamEntityAdapter::read($request);
     }
     public function create(Request $request)
     {
-        AbstractAdapter::create($request);
+        AbstractTeamEntityAdapter::create($request);
     }
 
     public function batchCreate(Request $request)
     {
-        AbstractAdapter::batchCreate($request);
+        AbstractTeamEntityAdapter::batchCreate($request);
     }
 
     public function update(Request $request)
     {
-        AbstractAdapter::batchCreate($request);
+        AbstractTeamEntityAdapter::batchCreate($request);
     }
 
     public function batchUpdate(Request $request)
     {
-        AbstractAdapter::batchUpdate($request);
+        AbstractTeamEntityAdapter::batchUpdate($request);
     }
 
     public function delete(Request $request)
     {
-        AbstractAdapter::delete($request);
+        AbstractTeamEntityAdapter::delete($request);
     }
 
     public function batchDelete(Request $request)
     {
-        AbstractAdapter::batchDelete($request);
+        AbstractTeamEntityAdapter::batchDelete($request);
     }
 
+    public function getMappedEntityClass()
+    {
+        return Site::class;
+    }
+
+    public function getMappedEntityName()
+    {
+        return 'site';
+    }
+
+    public function getMappedEntityDBName()
+    {
+        return 'site';
+    }
 }
